[FEAT] 로그아웃 api 개발(#S08P22D208-296), [FEAT] 유저 지갑주소 저장 api 개발(#S08P22D208-297)

See merge request s08-blockchain-contract-sub2/S08P22D208!53

+++ Server/module-api/src/main/java/com/example/kkaddak/api/dto/member/FollowResDto.java
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class FollowResDto {
-    private Long myFollowers;
-    private Long myFollwings;
+    private int myFollowers;
+    private int myFollwings;
 
     @Builder
-    public FollowResDto(Long myFollowers, Long myFollwings) {
+    public FollowResDto(int myFollowers, int myFollwings) {
         this.myFollowers = myFollowers;
         this.myFollwings = myFollwings;
     }
